another script to install ELK
The > /dev/null 2>&1 part of the command is used to redirect the standard output (1) and standard error (2) of the command to /dev/null. In other words, it discards any output from the command and sends it to a special file /dev/null which acts as a black hole for data, meaning anything sent to it is lost and can't be retrieved.

This redirection is used to prevent any output from the command being displayed in the terminal, making it useful for running commands that produce a lot of output that you don't need to see, or in cases where you don't want the output to clutter the terminal or log files.